fix(deep-zoom): zoom_at seam, f64 zoom storage, and high-precision correctness bundle (ARC-013, ARC-001, ARC-002/QA-001/002/004/005)
Phase 2a deep-zoom foundations (audit's infinite-zoom focus):
- ARC-013: extract single FractalParams::zoom_at() seam; replace the
  three triplicated zoom-at-cursor sites (shift+drag, scroll, pinch).
  Side effect: reconciling the pinch site to the reference formula fixes
  a latent 2x center-correction under-drift.
- ARC-001: zoom_2d f32 -> f64 (CPU-side + settings + attractor_last_zoom);
  GPU uniform stays f32, cast at the boundary; SEC-001 clamp updated to f64.
  Removes per-frame f32 rounding accumulation in zoom_2d *= factor.
- ARC-002 bundle (lands together per blocking relationship):
  - QA-002: burning_ship_hp DF abs negates both words via select (was
    injecting 2|lo| error/iter; sign(0) zeroed lo).
  - QA-005: Dekker-split two_prod (branch-free, correct without fused FMA);
    the FMA-based version silently collapsed DF to f32 on non-FMA backends.
  - QA-001: hp gate <=4u -> <=5u + explicit ==5u tricorn_hp branch; Tricorn
  - QA-004: hp auto-enable threshold 1e6 -> named HP_ZOOM_THRESHOLD=1e4
    (engages two decades earlier; derived pixel-spacing form deferred as it
    needs window_height not currently in Uniforms::update scope).

Uniform layout unchanged (864-byte assert holds; GPU zoom stays f32).
Visual verification (zoom sweep, Tricorn@1e6, BurningShip@1e8) pending Phase 4.

Verified: make checkall green (tests, clippy -D warnings, fmt).

+++ src/app/render.rs
@@ -118,7 +118,7 @@ impl App {
                     compute.uniforms = BuddhabrotComputeUniforms {
                         center_x: self.fractal_params.center_2d[0] as f32,
                         center_y: self.fractal_params.center_2d[1] as f32,
-                        zoom: self.fractal_params.zoom_2d,
+                        zoom: self.fractal_params.zoom_2d as f32,
                         aspect_ratio: self.renderer.size.width as f32
                             / self.renderer.size.height as f32,
                         width: self.renderer.size.width,
@@ -184,7 +184,7 @@ impl App {
                         param_d: 0.0,
                         center_x: self.fractal_params.center_2d[0] as f32,
                         center_y: self.fractal_params.center_2d[1] as f32,
-                        zoom: self.fractal_params.zoom_2d,
+                        zoom: self.fractal_params.zoom_2d as f32,
                         aspect_ratio: self.renderer.size.width as f32
                             / self.renderer.size.height as f32,
                         width: self.renderer.size.width,
